moving inventory models into a different file
